requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==2.1
django-extensions==2.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
idna==2.7
Pillow==5.2.0
PyJWT==1.6.4
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
urllib3==1.23
